backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
PyJWT==2.8.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.0.1
celery==5.3.6
redis==5.0.1
